fix(manager): PR #452 追加レビュー指摘に対応
- a=100/117(自前で<html>を出すアクション)を$isPaneRequestから除外し、
  フルページ応答へフォールバックさせる。footer.inc.phpの</main>出力も
  header.inc.php実行時のみに限定し、対応する<main>が無いまま閉じタグを
  出力する不整合を解消
- pane判定(isEvoPaneRequest)とfooter相当の後処理(evoRenderPaneFooterExtras)を
  helpers.phpの共通ヘルパーに集約し、index.php/accesscontrol.inc.php/
  footer.inc.php間の重複・ズレを解消
- header.inc.phpのJS文字列埋め込みをaddslashes()からjson_encode()へ変更
  (改行・</script>混入によるスクリプト破壊を防止)
- chromeless(QuickManager)表示でmainMenu不在時にdoRefreshが無限リトライ
  する既存不具合を修正
- shell.jsのPOST失敗時にフルリロードで復旧させ、無反応状態を解消
- resources_list.static.phpのpageパラメータをJS文字列へ埋め込む前に
  intキャスト(XSS対策、既存不具合)
- docmanager.class.phpのglobal $modx直接参照をconfig()ヘルパーに統一
- index.phpの$_POST/$_SESSION直接アクセスをpostv()/sessionv()ヘルパーに統一

+++ manager/actions/document/resources_list.static.php
@@ -385,7 +385,8 @@ function get_jscript($id, $cm)
 
     $contextm = $cm->getClientScriptObject();
     $textdir = $modx_textdir === 'rtl' ? '-190' : '';
-    $page = (getv('page')) ? " + '&page=" . getv('page') . "'" : '';
+    // JS文字列へ直接連結するため、数値以外を許容しない(XSS対策)
+    $page = (getv('page')) ? " + '&page=" . (int)getv('page') . "'" : '';
 
     $block = <<< EOT
 <style type="text/css">
